pyproject.toml
[tool.poetry]
name = "aadhaarlensapi"
version = "0.1.0"
description = "API for AadhaarLens app"
authors = ["Vishal Tanwar <vishal.tanwar@outlook.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
aadhaar-py = "^2.0.0"
Flask = "^2.0.2"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
flake8 = "^4.0.1"
add-trailing-comma = "^2.2.1"
isort = "^5.10.1"

[tool.isort]
profile = "black"
force_single_line = true

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
